refactor: body components to fit styled-components
