--- dikweb/components/body/index/IndexBody.module.css
@@ -1,19 +1,0 @@
-.indexBody {
-	display: flex;
-	margin: 28px 128px;
-}
-
-.indexBodyMidWrapper {
-	flex: 0.6;
-	margin: 18px 8px;
-}
-
-@media only screen and (max-width: 600px) {
-	.indexBody {
-		margin: 0px;
-	}
-	.indexBodyMidWrapper {
-		flex: 1;
-		margin: 0;
-	}
-}
